Implement append adif to make MGM easier

+++ mqt/MqtLogger/AdifImport.h
@@ -32,7 +32,7 @@ class ADIFImport
 
       LoggerContestLog * acontest;
       QSharedPointer<BaseContact> aqso;
-      int next_block;
+      unsigned long next_block;
    public:
       static bool doImportADIFLog( LoggerContestLog * c, QSharedPointer<QFile> hFile );
